webcall/pyproject.toml
[tool.poetry]
name = "webcall"
version = "0.1.0"
description = "WebRTC call backend with FastAPI (onion architecture)"
authors = ["You <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras=["standard"], version="^0.30.0"}
pydantic = "^2.8.2"
pydantic-settings = "^2.4.0"
SQLAlchemy = "^2.0.30"
asyncpg = "^0.29.0"
alembic = "^1.13.2"
redis = "^5.0.6"
passlib = {version = "^1.7.4", extras=["bcrypt"]}
python-jose = {version = "^3.3.0", extras=["cryptography"]}
httpx = "^0.27.0"
structlog = "^24.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-asyncio = "^0.23.8"
httpx = {version="^0.27.0", extras=["http2"]}
ruff = "^0.6.3"
black = "^24.8.0"
isort = "^5.13.2"
pre-commit = "^3.8.0"
mypy = "^1.11.1"

[tool.ruff]
line-length = 100
extend-select = ["I"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
filterwarnings = [
  "ignore::DeprecationWarning",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
